requirements.txt
flask
pymongo
requests
google-generativeai
python-dotenv
